requirements.txt
Flask==1.1.1
flask-ngrok==0.0.25
fpdf==1.7.2
gunicorn==20.0.0
image==1.5.27
imageio==2.4.1
imagesize==1.1.0
itsdangerous==1.1.0
kaggle==1.5.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vis==0.4.1
matplotlib==3.1.1
matplotlib-venn==0.11.5
nltk==3.4.5
np-utils==0.5.11.1
numpy==1.17.4
pandas==0.25.3
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.2
semantic-version==2.8.2
sklearn==0.0
sklearn-pandas==1.8.0
sqlparse==0.3.0
tblib==1.5.0
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-datasets==1.3.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.7.0
tensorflow-metadata==0.15.1
tensorflow-privacy==0.2.2
tensorflow-probability==0.7.0
tflearn==0.3.2
Theano==1.0.4
tornado==4.5.3
tqdm==4.28.1
urllib3==1.24.3
wasabi==0.4.0
xlrd==1.1.0
xlwt==1.3.0
Pillow==6.2.0
